double only confuses

+++ test.c
@@ -57,7 +57,7 @@ static define_effect_fn(half_mad, struct HalfTernary const *arg) {
 }
 
 static void test_premul(void) {
-    struct Color c = {0.75, 0.5, 0.25, 0.5};
+    struct Color c = {0.75f, 0.5f, 0.25f, 0.5f};
     struct Registers reg;
 
     struct Effect program[] = {
@@ -69,10 +69,10 @@ static void test_premul(void) {
     run(program,1);
 
     for (int i = 0; i < len(reg.r); i++) {
-        expect((double)reg.r[i] == 0.375);
-        expect((double)reg.g[i] == 0.250);
-        expect((double)reg.b[i] == 0.125);
-        expect((double)reg.a[i] == 0.500);
+        expect((float)reg.r[i] == 0.375f);
+        expect((float)reg.g[i] == 0.250f);
+        expect((float)reg.b[i] == 0.125f);
+        expect((float)reg.a[i] == 0.500f);
     }
 }
